Microsoft.CSharp: Remove check for ExprMemberGroup being used as a lvalue (#25214)
* Remove check for ExprMemberGroup being used as a lvalue.

ExprMemberGroups only exist within ExprCall objects and can never be
passed as operands, so this is unreachable. (Unlike in static
compilation where it would be the result of someone trying e.g.
obj.ToString = 3).

Entails removal of ERR_AssgReadonlyLocalCause, contributes to #22470

* Remove ErrArgIds class

Only place it was instantiated has now been removed.

Clean-up all unused ErrArgKind as well as Ids that this makes unused.