Require explicit action to read a non-universal tag
For each method on the reader add an overload which takes the expected tag.
The actual tag and expected tag have to match on class and value, but not the
primitive/constructed bit.

It is not allowed to "expect" the wrong universal tag.

The overload without the expectedTag parameter will delegate with the
appropriate universal tag for the method.