Harden DerSequenceReader against unsupported data.
Because our data is bounded within a .NET array it cannot exceed 2GB, so
anything that looks to be larger is a read error.

Additionally, multi-byte tags are fairly rare, and we would process one as a one
byte tag and a corrupt length, so detect that we're about to go invalid and stop
early.