Place lambda's return expression directly inside ExprBoundLambda
Since the expression is always a single block with a single return that
returns a single expression, just store that single expression.

Essentially this changes the restrictions from what one can do in C# in
defining an anonymous delegate to what one can do in C# in defining an.

It matches what we are actually using the object for better, and
removes two levels of indirection through other objects, with the
possibility of errors in them.