Speed up Windows ProcessManager.OpenProcess
If a process is running as a normal user (not elevated), many calls to
`kernel32.OpenProcess` fail.

In `ProcessManager.OpenProcess`, when the first call to `kernel32.OpenProcess`
fails, another call is made indirectly in `ProcessManager.IsProcessRunning`
which will also fail.

When `ProcessManager.IsProcessRunning` can't get a handle to the process, it
falls back on a much more expensive check - enumerating all processes.

The fix is to assume the process is running if the error code is
ERROR_ACCESS_DENIED.

This performance issue adds roughly 1 second of overhead when running
`Get-Process` in an interactive non-elevated PowerShell session where
the CPU property is displayed.