Replace instead of push to stack in S.L.Expressions interpreter.
The interpreter does a lot of stack pushes and pops, but very often
pushing right after a pop, and often pushing the value just popped.

Offering a Replace method means that not only can a Peek/Replace avoid
changing the stack pointer, but often it can leave the stack alone, or
short-circuit a lifted operation after examining only one operand.

Do so.