Remove isVarargs and isAbstract from MethodSymbol
isVarargs had only been used in error messages (no longer the case) and
an assertion (can be replaced with a check for an ArgumentListType
parameter), with varargs members correctly failing overload resolution
due to the ArgumentListType parameter not matching rather than the
isVarargs flag.

isAbstract was set but never used.