Add a define-based block to guard each method from write-ahead
Since EnsureWriteCapacity and ArrayPool both have some logic to round
up to a "more useful" number individual methods in the writer could end
up writing to _buffer at an offset higher than they have guaranteed, and
it will almost always succeed.  With CHECK_ACCURATE_ENSURE the buffer is
never bigger than it "needs to be" so bad math on ensure will result in an
ArrayIndexOutOfBoundsException, or other fault.