Simplify some casting and conversion related code in Microsoft.CSharp (#25357)
* Remove pDestinationTypeForLambdaErrorReporting from conversions.

Stored in a field and passed around, but never used for anything beside

* Don't use ExprClass in casts and typeof

Acts as a wrapper around the type, but provides nothing beyond that
wrapper, so it is just unwrapped again. (Visiting an ExprClass is
always a nop, so that's not a reason to wrap it).

Use the CType directly instead.

* Make visiting ExprClass explicitly just return it directly.

This is what happens anyway, but since the previous change depends on
this, making this clear makes that being correct clearer, as well as a
minor efficiency gain.

* Make ExprTypeOf.SourceType init-only.

As never changed now.