Add basic Socket.IOControl implementation on Unix
Socket.IOControl uses Windows-specific code values, both in terms of the actual code numbers and in terms of lots of them being Windows-specific functionality.  But a few of the codes do translate across OSes in terms of functionality, and we've heard from at least a few developers who were using these codes and expected them to work on Unix.