Dependably determine collectibility in S.L.Expression's CanCache
Use RuntimeType.IsCollectible() to determine collectibility of types
rather than only assuming those in mscorlib are not, which has many
false positives (over 99% false positives in the uses Microsoft.CSharp
makes of expressions).

Build a delegate to make the call when IL generation is available, or
use reflection directly when it is not.