Fix Alpine CURL concurrency issue
On Alpine Linux, CURL uses LibreSSL for its HTTPS needs. That creates
the same concurrency issue that we had with LibreSSL on OSX. It was
fixed for OSX only in the past by detecting which SSL library CURL uses
and using shared MultiAgent for all CurlHandler instances.
This change extends it to all Unixes, since if CURL uses LibreSSL on a
Linux distro, we need the same workaround.