Remove dead code from MS.CSharp's BadOperatorTypesError and related methods (#25711)
* Remove dead paths in BadOperatorTypesError

pOperand is never null, and neither operand can ever have a null or
ErrorType type, so remove path for those possibilities.

* Remove unused parameters from BadOperatorTypesError

ek and pTypeErr are never used, so remove them. Merges the overloads.

* Change BadOperatorTypesError to return rather than throwing exceptions

Since it throws in every path, have the error thrown at the call site,
making it more clearly an error path.

* Remove UnwrapExpression

The only place ExprWraps are produced outside of the
ExpressionTreeRewriter is in BindUserBoolOp which passes them to
bindUDUnop. There's no path that can bring them to
BadOperatorTypesError which calls UnwrapExpression.

* Have CalculateExprAndUnaryOpKinds return a value tuple and throw ICE on
failure.

Rather than have an impossible condition checked for outside of it.
(ICE is more appropriate, too).

* Have GetBinopKindAndFlags return value tuple and throw ICE on failure.

Rather than have an impossible condition checked for outside of it.