Create a Memory/Span-based ASN.1 reader and writer (#25296)
The AsnReader type can read any BER-encoded value for the types that it supports (for tag values up to int.MaxValue). It also understands the CER and DER restrictions, and when reading in those modes it will enforce the restrictions that those encoding rulesets dictate.  Callers who want minimal validation can just read in BER mode.

The AsnWriter type mostly writes in DER (except where CER demands a different encoding). The only significant "BER-relaxation" it takes is that closing a SET OF value will not sort the contents.

Reference materials:

* ITU-T-REC-X.680-201508 (ASN.1 language, and some semantics behind the values)
* ITU-T-REC-X.690-201508 (the BER encoding family (BER, CER, DER))