Hotfix for the Alpine CI issue (#25806)
The Alpine CI is failing due to an old version of dotnet host package
specified in the dependencies.props file. There are no Alpine packages
for that version. However, updating that version causes Invariant.Tests
tests to fail for an unclear reason.
So I am making this change that disables building tests for Alpine in
the CI. That means we will still build both native and managed code of
corefx. This change is temporary until the Invariant test issue is
understood and fixed.