NamedPipeClient connect throws TimeoutException when appropriate
Fixes #24635

Changing the code to ERROR_SEM_TIMEOUT changes the exception from
IOException("The semaphore timeout period has expired") to a
TimeoutException when the pipe is busy.

Another change in behavior: if there was a good cancellationToken (i.e.:
CanBeCancelled was true) with a timeout larger than 50ms and the pipe
was busy it was failing with the IOException on first try (after waiting
50ms). However it was waiting for the whole user timeout if the token
could not be cancelled. This change will make it wait for the full user
timeout independently of cancellationToken.CanBeCancelled value.