MS.CSharp Refactoring of producing lambda for dynamic operation and rewriting it (#25390)
* Tighten type of parents in symbol factory methods to those valid.

* Remove parent parameter to SymFactory.CreateScope()

Always null; we only ever work with one scope. Let's make that explicit

* Remove RewriteLambdaParameters

Pulls locals that are sibling to scope into lambda expression (for
hoisting) but there is never a sibling to any scope in the dynamic
binder.

* Remove ExpressionTreeRewriter.DestroyWraps

Unwraps hoisted sibling locals, but we never have any.

* Remove ExpressionTreeRewriter.currentAnonMeth

We only ever have one bound lambda at a time; never another nested
within that, so this check is unnecessary.

* Remove check createParameters != null

Always true. Indeed ExpressionTreeCallRewriter.Rewrite depends on this
as it expects an ExprBinOp, and if this path wasn't taken we'd return
an ExprCall instead.

* Take VisitBOUNDLAMBDA out of the main visiting loop.

Can only be reached from the top of an ExpressionTreeRewriter visit, so
only call it from there, tightening parameter types.

* Tighten VisitBoundLambda to return ExprBinOp

Pass this tightened type definition through to
ExpressionTreeCallRewriter.Rewrite, making one asserted fact there
statically known.

* Remove ExpressionKind.SequenceReverse

No such expression is ever created.

* Create predefined type for Func<>

Since we're using it with every dynamic operation, it should be in the
faster lookup.

* Tighten ExprBoundLambda ctor parameter types and assertions.

* Place lambda's return expression directly inside ExprBoundLambda

Since the expression is always a single block with a single return that
returns a single expression, just store that single expression.

Essentially this changes the restrictions from what one can do in C# in
defining an anonymous delegate to what one can do in C# in defining an.

It matches what we are actually using the object for better, and
removes two levels of indirection through other objects, with the
possibility of errors in them.

* Remove ExprBlock, ExprReturn & ExprStatement

No longer used.

Remove commented-out ExpressionKinds along with those removed here.

* Don't add "this" local to scope.

Never given meaning, only ever deliberately avoided in rewrites.

When a dynamic operation uses this, it does so by having this passed-in
explicitly.

* Remove LocalVariableSymbol.fUsedInAnonMeth

Never not set to true, so no real significance.

* Pass deferred binding to out parameter directly.

We'll always get the correct final result there, so don't replicate
logic we just called into.

* Remove path in ExpressionTreeRewriter.VisitLOCAL for null wrap.

Since the only locals have just had this set to non-null, this is
unreachable.

Since the code removed here is the only use of ExprHoistedLocalExpr,
remove it too.

* Remove obsolete comment.