Update cng and openssl packages to pin netstandard assembly versions
Given that we are harvesting the netcoreapp2.0 assets for these packages
we need to keep the netstandard2.0 assembly versions matching what we are harvesting.

If we need to add more APIs to the netstandard refs then we need need to bump the versions
and start building live versions of the netcoreapp2.0 assets again to add the APIs.