Allow RangeAttribute use to request InvariantCulture conversions.
Adds a ParseLimitsInInvariantCulture that affects the interpretation of
the extrema set, and ConvertValueInInvariantCulture that affects the
interpretation of values.

Separate methods allows for e.g. a component assuming a
dot-decimal-separator culture defining a property with the attribute
(so invariant culture should be used) the use of the property is done
in a comma-decimal-separator (so the current culture should be used),
or any permutation of the two booleans.

Fixes #2648