Freeze the assembly version for System.Threading.Tasks.Extensions
The APIs in System.Threading.Tasks.Extensions have been put inbox on
.NET Core 2.0 and UAp and also type-forwarded into System.Runtime. This
means the APIs in the contract can no longer version for the netstandard
reference. Any new API's have to be added to the platform specific reference
assembly.

This also updates the package to include the inbox placeholders for
netcoreapp and uap.