Add RuntimeFeature.DefaultImplementationsOfInterfaces
The flag indicates that the runtime supports default implementation of interfaces.

Since this is a prerelease feature that is not complete yet (and we don't want to accidentally ship it), the flag is keyed off a newly added `IsPrerelease` property. This property is expected to be flipped to `false` in release branches.

This corresponds to dotnet/coreclr#15358.