Delete SignSilentDoesNotPrompt test
This test adds between 6 and 12 seconds of execution on my machine,
depending on the mood of my smart card drivers, including at least one run
where a smart card driver never answered the query for
CryptAcquireCertificatePrivateKey (or, at least, not within 40+ seconds).

Without this test, execution time drops to 0.6 seconds for the suite.