Ensure System.Uri consistently uses the correct RFC 3986 reserved character set (#26006)
Resolve an issue with System.Uri where certain percent encoded characters that should be left as-is are decoded, but only when the URI also contains Unicode or an encoded non-reserved symbol. 

This issue was left over after the URI reserved character set was upgraded from RFC 2396 to RFC 3986. In several places the upgrade was missed, resulting in certain code paths still using the old reserved character set. 

This fix updates places where the RFC 2396 set was still being used, and adds tests to ensure this behavior is consistent. It also retains the code used for quirking in .Net Framework, but always enables the new behavior instead of checking an AppContextSwitch. 

This fix has been ported from .NET Framework.