Only set Headers.ContentLength if it isn't already correct
An unmanaged COM exception is thrown by Windows.Web.Http.dll when an
attempt to set an already-present header is made. This particular line
is in the hot path and is almost always invoked on outgoing requests
causing these COM exceptions to be thrown all the time. Avoiding
downstream exceptions should make things faster and result in fewer
pipeline flushes, etc.