Add Brotli to existing native Compression builds
Because of some requirements for crc functions (log and exit), I had to open up the Windows AOT release native clrcompression builds to import from the crt api sets. My understanding is that this is fine as we are now packaging the crt api sets alongside microsoft.netcore.runtime.coreclr so we don't have to worry about them missing on the target machine.

Similarly, the Unix native build had to change slightly as the current build makes use of some stringent checking on the target code which makes sense since they're currently just shims. I moved the failing checks to only be imposed on the still-shim native builds and ignore System.IO.Compression which is now filled with Brotli code.