Make AsnWriter IDisposable
Without this change the AsnWriter class always sends the last "rented" array to
the garbage collector.  Since arrays from the array pool are expected to be
gen-2 highly reused objects this has unfortunate consequences for both
the induced trickle allocation and Gen2 pressure.

Now it's IDisposable, and the Dispose implementation will return the rented array.