Fix Microsoft.Diagnostics.Tracing.EventSource.Redist Mirroring
We used to mirror some files from this library to some shared folder from coreclr repo. Now we got the mirroring working for the whole corelib shared folder to corefx. So, this change is getting rid of the shared redundent files in this library and use only the mirrored shared folder.