Add another Zip IEnumerable<T> extension method
There's already a Zip extension method that takes two IEnumerables
and a result selector Func that's meant to combine the elements
of the enumerables in some way. Perhaps the most common use of this
method is to just create a tuple, so I've added an overload which
removes the result selector and instead produces a ValueTuple by
default.