Port ECDiffieHellman to .NET Core
This brings the ECDiffieHellman class as-is from .NET Framework (last edited
for net462) to System.Security.Cryptography.Algorithms, and the
ECDiffieHellmanCng class as-is to System.Security.Cryptography.Cng, and adds
ECDiffieHellmanOpenSsl to System.Security.Cryptography.OpenSsl.

A fair amount of this change is factoring out ECDSA code to general ECC code
for each of the platforms.

Since there isn't obvious value in having public ECDiffieHellmanPublicKey
derived types, ECDiffieHellmanOpenSslPublicKey is not public.