Refactor MS.CSharp's CType (#26414)
* Remove name from CType.

Only used in TypeParameterType, and there can be done by delegating to
symbol.

* Make CType.Get/SetTypeKind init-only property.

* Have singleton CTypes control their own singleton instances.

* Initial AggregateType state as init-only properties

* TypeParameterType state to init-only property.

All other getter methods to properties.

* ArrayType state as auto-properties.

Other getter method as property.

* PointerType state to init-only auto-properties.

* ParameterModifierType state to init-only auto properties

* NullableType state to init-only auto properties.

* Remove TypeFactory

No state or returning of derived types to benefit from factory pattern.

* Don't use names in TypeTable lookups.

The names are created to reflect properties of the type, but just use
the values of those properties directly.

* Don't branch on result of TryGet within TypeTable

In case of failure out parameter is guaranteed to be null, so no need
to branch just to return null.

* Replace getAggregate() calls with OwningAggregate where feasible.

Avoid cast, or move cast closer to where it is deemed safe. Greater
type-safety.

* CType IsRefType() to virtual property.

* IsNonNubValType() and IsValType() to virtual properties

* IsStaticClass() to virtual member.

Note that while in source code an array, parameter modifier or pointer
can be of a static type due to an error in the source, these types can
never be static in dynamic code.

* Remove arguments to CheckForStaticClass that are always the same.

* isPredefined and getPredefType to virtual properties.

* isPredefType() to virtual instead of type-testing.

* isUnsafe to virtual instead of type-testing.

* Several Aggregate-only methods to virtual properties.

* Move IsCollectionType to only derived type that uses it.

Optimise to skip tests it can't pass.

* Make AssociatedSystemType virtual instead of testing types.

Only memoise in AggregateType as it's not as expensive elsewhere.

* Make GetBaseOrParameterOrElementType virtual property.

* IsWindowsRuntimeType to property.

* Remove underlyingType()

All uses can be safely replaced with UnderlyingEnumType and name
clashes with UnderlyingType property elsewhere.

* Remove more uses of getAggregate()

One is within a branch for a test that is only true for AggregateTypes.
Another goes from an enum to the underlying before using
Type.GetTypeCode() on the system type, but Type.GetTypeCode() will
behave the same on the enum type anyway.

* IsUnsigned to virtual property.

* fundType() to virtual property

* constValKind to virtual property.

* AggregateType.GetIfacesAll to property.

* AggregateType.GetBaseClass to property

* Remove CheckConstraintsFlags.NoDupErrors

Checked for, but never set.

* Merge AggregateType.fConstraintsChecked and fConstraintError into bool?

* Don't recheck constraints if exception not necessary.

Akin to the removed NoDupErrors but where that prevented repeated
messages in a static compilation, with dynamic code we can depend on
the cached value from previous operations.

* Change field names to not use Hungarian notation.

* Short-circuit on constraints checking.

The non–short-circuiting behaviour is a legacy of the times the static
compiler would want to emit multiple error messages. The dynamic
compiler will already short-circuit for error-emitting cases as the
exception will cut the operation short, but it should also short-
circuit for non–error-emitting cases too.

* Pick nit.