Replace TryCopyTo with CopyTo
While doing various Span-related work involving copying, in a few places I previously used a pattern of calling TryCopyTo and then asserting its result, in places where I knew the destination was long enough. This was because CopyTo was implemented as a wrapper around TryCopyTo and thus involved an extra unnecessary branch. Now that that's no longer the case, I'm simplifying the call sites.

cc: @GrabYourPitchforks