Utf8Parsing - the remaining 'N' format overloads
Fixes https://github.com/dotnet/corefx/issues/24986

This is a piece of the Utf8Parser that was punted for time
last year.

There are 8 overloads for each of the 8 integer types.
Int32 was done in a previous PR - this completes the set.

The "N" format prints out integers like this:

  `"N2" => 12,345.00`

This parser mimics the behavior of

 `int.TryParse(v, NumberStyles.Integer | AllowThousands | AllowDecimalPoint)`

The thing that may look strange is that the parser
allows commas anywhere, not just on the 10^3 digits.
This mimics the desktop api behavior. Comma placement is
culture-dependent and this is an api that is culture-agnostic.

The test data was confirmed on a control implementation
that calls the classic int.TryParse().