Make EnvelopedCms work for Linux and macOS
This change adds a version of EnvelopedCms which uses the platform crypto
stack and new AsnSerializer capabilities, and makes no P/Invokes (other than
the ones the crypto stack itself makes).

.NET Framework EnvelopedCms is more or less RFC 2630 (CMS version "2")
plus RFC 3565 (AES with CMS), and that is maintained with this implementation.
No new API has been added, this merely populates the existing API for
non-Windows OSes.

Unlike the Windows implementation, FFC based Diffie-Hellman key
agreement is not supported, because the platform has no exposure of FFC
Diffie-Hellman.

Unlike the WIndows implementation, RC4 is not supported, because the
platform has no exposure of RC4.