Fix ServicePointManager.ProxyAddressIfNecessary to ignore "system" proxy failures (#26925)
The whole ServicePointManager implementation is there just to make basic stuff not fail.  But HttpWebRequest's Proxy defaults to an internal dummy singleton SystemWebProxy whose GetProxy method throws a PlatformNotSupportedException.  Until we can clean that up and have a real proxy in place, we should just eat that PlatformNotSupportedException when the ServicePointManager calls GetProxy.