Fix ProcessStartInfo tests on CentOS in Debug
There is a bug in CentOS7.1 getpwnam_r where it returns the wrong value if you pass in a buffer that is too small to hold the data.  In Debug builds, we are purposefully passing in a buffer that is too small to test the fallback code, which is running into this OS bug.

See https://bugs.centos.org/view.php?id=7324

Fixes #26673